+++ src/main/kotlin/icu/windea/pls/lang/settings/PlsSettings.kt
@@ -11,9 +11,7 @@ import icu.windea.pls.lang.util.*
 import icu.windea.pls.model.*
 
 /**
- * PLS设置。
- *
- * 可以在插件的设置页面（`Settings > Languages & Frameworks > Paradox Language Support`）中进行配置。
+ * PLS设置。可以在插件的对应设置页面中进行配置。
  */
 @Service(Service.Level.APP)
 @State(name = "ParadoxSettings", storages = [Storage(PlsConstants.pluginSettingsFileName)])